websockets/pyproject.toml
[tool.poetry]
name = "websockets"
version = "0.1.0"
description = ""
authors = ["luikymagno <luikymagno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
websocket-client = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
